rename package to "shorthand"
