+++ setup.py
@@ -3,7 +3,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='note_parser',
+    name='shorthand',
     version='0.1',
     description='Note Management utilities',
     packages=find_packages()
