+++ shorthand/utils/config.py
