+++ shorthand/utils/logging.py
