+++ shorthand/utils/patterns.py
