+++ shorthand/utils/typeahead.py
