+++ web/api.py
@@ -12,37 +12,37 @@ from datetime import datetime
 
 from flask import Flask, request, render_template, send_from_directory
 
-from note_parser.todo_tools import get_todos, mark_todo, stamp_notes
-from note_parser.search_tools import search_notes, get_context, get_note
-from note_parser.question_tools import get_questions
-from note_parser.tag_tools import get_tags
-from note_parser.calendar_tools import get_calendar
-from note_parser.toc_tools import get_toc
-from note_parser.utils.config import get_notes_config
-from note_parser.utils.logging import setup_logging
-from note_parser.utils.render import get_file_content, get_rendered_markdown
-from note_parser.utils.typeahead import get_typeahead_suggestions
+from shorthand.todo_tools import get_todos, mark_todo, stamp_notes
+from shorthand.search_tools import search_notes, get_context, get_note
+from shorthand.question_tools import get_questions
+from shorthand.tag_tools import get_tags
+from shorthand.calendar_tools import get_calendar
+from shorthand.toc_tools import get_toc
+from shorthand.utils.config import get_notes_config
+from shorthand.utils.logging import setup_logging
+from shorthand.utils.render import get_file_content, get_rendered_markdown
+from shorthand.utils.typeahead import get_typeahead_suggestions
 
 
 app = Flask(__name__)
 
-NOTES_CONFIG = get_notes_config()
-log = setup_logging(NOTES_CONFIG)
+SHORTHAND_CONFIG = get_notes_config()
+log = setup_logging(SHORTHAND_CONFIG)
 
 
 @app.route('/', methods=['GET'])
 def show_ui():
 
     all_directories = ['ALL']
-    for subdir in os.walk(NOTES_CONFIG['notes_directory']):
-        subdir_path = subdir[0][len(NOTES_CONFIG['notes_directory']) + 1:]
+    for subdir in os.walk(SHORTHAND_CONFIG['notes_directory']):
+        subdir_path = subdir[0][len(SHORTHAND_CONFIG['notes_directory']) + 1:]
         if '.git' in subdir_path or not subdir_path:
             continue
         elif len(subdir_path.split('/')) > 2:
             continue
         else:
             all_directories.append(subdir_path)
-    default_directory = NOTES_CONFIG.get('default_directory')
+    default_directory = SHORTHAND_CONFIG.get('default_directory')
 
     log.warning('Showing the Home page!')
     return render_template('index.j2', all_directories=all_directories,
@@ -70,7 +70,7 @@ def get_current_todos():
         directory_filter = None
 
     return json.dumps(get_todos(
-                notes_directory=NOTES_CONFIG['notes_directory'],
+                notes_directory=SHORTHAND_CONFIG['notes_directory'],
                 todo_status=status, directory_filter=directory_filter,
                 query_string=query_string, sort_by=sort_by, suppress_future=True))
 
@@ -84,7 +84,7 @@ def fetch_questions():
         directory_filter = None
 
     return json.dumps(get_questions(
-                notes_directory=NOTES_CONFIG['notes_directory'],
+                notes_directory=SHORTHAND_CONFIG['notes_directory'],
                 question_status=status, directory_filter=directory_filter))
 
 
@@ -96,7 +96,7 @@ def fetch_tags():
         directory_filter = None
 
     return json.dumps(get_tags(
-                notes_directory=NOTES_CONFIG['notes_directory'],
+                notes_directory=SHORTHAND_CONFIG['notes_directory'],
                 directory_filter=directory_filter))
 
 
@@ -108,7 +108,7 @@ def fetch_calendar():
         directory_filter = None
 
     return json.dumps(get_calendar(
-                notes_directory=NOTES_CONFIG['notes_directory'],
+                notes_directory=SHORTHAND_CONFIG['notes_directory'],
                 directory_filter=directory_filter))
 
 
@@ -119,7 +119,7 @@ def get_search_results():
     case_sensitive = request.args.get('case_sensitive')
 
     return json.dumps(search_notes(
-                notes_directory=NOTES_CONFIG['notes_directory'],
+                notes_directory=SHORTHAND_CONFIG['notes_directory'],
                 query_string=query_string,
                 case_sensitive=case_sensitive))
 
@@ -132,8 +132,8 @@ def get_line_context():
     width = int(request.args.get('width', 5))
 
     # Allow Relative paths within notes dir to be specified
-    if NOTES_CONFIG['notes_directory'] not in filename:
-        filename = NOTES_CONFIG['notes_directory'] + filename
+    if SHORTHAND_CONFIG['notes_directory'] not in filename:
+        filename = SHORTHAND_CONFIG['notes_directory'] + filename
 
     return json.dumps(get_context(filename, line_number, width))
 
@@ -143,8 +143,8 @@ def get_full_note():
 
     path = request.args.get('path')
 
-    if NOTES_CONFIG['notes_directory'] not in path:
-        path = NOTES_CONFIG['notes_directory'] + path
+    if SHORTHAND_CONFIG['notes_directory'] not in path:
+        path = SHORTHAND_CONFIG['notes_directory'] + path
 
     return get_note(path)
 
@@ -152,7 +152,7 @@ def get_full_note():
 @app.route('/render', methods=['GET'])
 def send_rendered_note():
 
-    file_path = NOTES_CONFIG['notes_directory'] + request.args.get('path')
+    file_path = SHORTHAND_CONFIG['notes_directory'] + request.args.get('path')
     file_content = get_file_content(file_path)
     file_content = get_rendered_markdown(file_content)
     file_content = file_content.replace("\\", "\\\\")
@@ -164,7 +164,7 @@ def send_rendered_note():
 @app.route('/calendar', methods=['GET'])
 def show_calendar():
 
-    summary = get_calendar(NOTES_CONFIG['notes_directory'])
+    summary = get_calendar(SHORTHAND_CONFIG['notes_directory'])
     events = []
     for year, year_data in summary.items():
         for month, month_data in year_data.items():
@@ -182,7 +182,7 @@ def show_calendar():
 @app.route('/chart', methods=['GET'])
 def show_chart():
 
-    summary = get_calendar(NOTES_CONFIG['notes_directory'])
+    summary = get_calendar(SHORTHAND_CONFIG['notes_directory'])
     timeline_data = []
     for year, year_data in summary.items():
         for month, month_data in year_data.items():
@@ -200,12 +200,12 @@ def show_chart():
 
 @app.route('/toc', methods=['GET'])
 def get_toc_data():
-    return json.dumps(get_toc(NOTES_CONFIG['notes_directory']))
+    return json.dumps(get_toc(SHORTHAND_CONFIG['notes_directory']))
 
 
 @app.route('/browse', methods=['GET'])
 def show_browse_page():
-    toc = get_toc(NOTES_CONFIG['notes_directory'])
+    toc = get_toc(SHORTHAND_CONFIG['notes_directory'])
     return render_template('browse.j2', toc=json.dumps(toc))
 
 
@@ -217,8 +217,8 @@ def mark_todo_status():
     status = request.args.get('status')
 
     # Allow Relative paths within notes dir to be specified
-    if NOTES_CONFIG['notes_directory'] not in filename:
-        filename = NOTES_CONFIG['notes_directory'] + filename
+    if SHORTHAND_CONFIG['notes_directory'] not in filename:
+        filename = SHORTHAND_CONFIG['notes_directory'] + filename
 
     return mark_todo(filename, line_number, status)
 
@@ -229,14 +229,14 @@ def get_typeahead():
     query_string = request.args.get('query')
 
     return json.dumps(get_typeahead_suggestions(
-        NOTES_CONFIG['ngram_db_directory'],
+        SHORTHAND_CONFIG['ngram_db_directory'],
         query_string))
 
 
 @app.route('/stamp', methods=['GET'])
 def stamp():
 
-    return stamp_notes(NOTES_CONFIG['notes_directory'])
+    return stamp_notes(SHORTHAND_CONFIG['notes_directory'])
 
 
 if __name__ == "__main__":
